requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argcomplete==3.5.3
asgiref==3.8.1
asttokens==3.0.0
attr==0.3.1
attrs==24.3.0
azure-core==1.32.0
azure-storage-blob==12.24.0
bcrypt==4.3.0
black==24.10.0
bleach==5.0.1
blinker==1.9.0
boto==2.49.0
boto3==1.35.97
botocore==1.35.97
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
datamodel-code-generator==0.26.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
Django==4.2.17
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==2.4.0
django-migration-linter==5.1.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dnspython==2.7.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email_validator==2.2.0
executing==2.2.0
expiringdict==1.2.2
Faker==33.3.1
filelock==3.17.0
fonttools==4.55.4
fsspec==2024.12.0
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.161.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.4
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
inflect==5.6.2
inflection==0.5.1
ipykernel==6.29.5
ipython==8.31.0
isodate==0.7.2
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
label-studio==1.15.0
label-studio-sdk==1.0.8
langchain-core==0.3.41
langchain-ollama==0.2.3
langsmith==0.3.11
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
narwhals==1.34.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
ollama==0.4.7
openai==1.59.7
opencv-python==4.11.0.86
opentelemetry-api==1.29.0
ordered-set==4.0.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyboxen==1.3.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
PyPDF2==3.0.1
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytz==2022.7.1
pywin32==308; platform_system == "Windows"
PyYAML==6.0.2
pyzmq==26.2.0
redis==3.5.3
referencing==0.35.1
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rq==1.10.1
rsa==4.9
rstr==3.2.2
rules==3.4
s3transfer==0.10.4
safetensors==0.5.3
scipy==1.15.1
sympy==1.13.3
seaborn==0.13.2
semver==3.0.2
sentry-sdk==2.19.2
setuptools==75.8.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.44.1
tenacity==9.0.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchvision>=0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.0
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
ultralytics==8.3.65
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.40.0
wrapt==1.17.1
xmljson==0.2.1
zipp==3.21.0
zstandard==0.23.0
